Tien-Thu comit
ALL

+++ backend_fn/database/migrations/2024_11_13_173758_create_payment_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('payment', function (Blueprint $table) {
             $table->id();
-            $table->decimal('payment_amount');
+            $table->integer('payment_amount');
             $table->dateTime('payment_date');
             $table->string('payment_method');
             $table->string('payment_status')->default(1);
